refactor: move login/logout/etc. cmds to cfg ns
