+++ commands/cfg/location.go
@@ -1,0 +1,28 @@
+package cfg
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/ionos-cloud/ionosctl/v6/pkg/config"
+	"github.com/ionos-cloud/ionosctl/v6/pkg/core"
+)
+
+func CfgLocationCmd() *core.Command {
+	cmd := core.NewCommand(context.Background(), nil, core.CommandBuilder{
+		Namespace: "config",
+		Resource:  "location",
+		Verb:      "location",
+		Aliases:   []string{"location", "loc"},
+		ShortDesc: "Print your config file's path",
+		Example:   "ionosctl cfg loc",
+		PreCmdRun: core.NoPreRun,
+		CmdRun: func(c *core.CommandConfig) error {
+			_, err := fmt.Fprintf(c.Command.Command.OutOrStdout(), config.GetConfigFile())
+			return err
+		},
+		InitClient: false,
+	})
+
+	return cmd
+}
