+++ commands/root.go
@@ -160,10 +160,9 @@ func initConfig() {
 // AddCommands adds sub commands to the base command.
 func addCommands() {
 	rootCmd.AddCommand(VersionCmd())
-	rootCmd.AddCommand(cfg.LoginCmd())
-	rootCmd.AddCommand(cfg.LogoutCmd())
-	rootCmd.AddCommand(cfg.CfgLocationCmd())
-	rootCmd.AddCommand(cfg.WhoamiCmd())
+	rootCmd.AddCommand(cfg.LoginCmd()) // Hidden. Users will only find out help about the one in `config` namespace
+	rootCmd.AddCommand(cfg.ConfigCmd())
+
 	// V6 Resources Commands
 	rootCmd.AddCommand(cloudapiv6.LocationCmd())
 	rootCmd.AddCommand(cloudapiv6.DatacenterCmd())
